Solution to Perform Bitwise One's Complement

+++ Perform Bitwise One's Complement@@67882a731b4392d8fc1577e7/code.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 int main() {
-    int num,result;
+    long int num,result;
     scanf("&d",&num);
     result=~num;
     printf("%d\n",result);
